pyproject.toml
[tool.poetry]
name = "emergence-bot"
version = "1.0.0"
description = "An all-purpose discord bot for personal use."
authors = ["CephieCodesStuff <yourlittlecephalon@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
discord-py = {extras = ["voice"], version = "^2.1.0"}
python-dotenv = "^0.21.0"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
black = "^22.10.0"
aiohttp = "^3.8.6"
aiodns = "^3.1.1"
rapidfuzz = "^3.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
